Updated jetson setup -- pyenv python2.7 and python3; install opencv under ~/.local
Jetson setup now properly creates virtual python environments for 2.7 and 3 and pip
installs everything into the virtual environments.

The custom build of opencv is now installed into ~/.local. Thios prevents it from colliding
with apt-get installed opencv. It is earlier in the path so it takes priority.

import cv2 works in both python2.7 and python3.